Dashboard with checkbox
Hey Alex, 

Here is the file I was talking about, if you could move around the text to the appropriate places that would be awesome! Also anything that you can do to make it look "sleek" would be awesome. Pretty bummed that we lost the toggle but will have to roll that out with v2 of the app I guess.